Major update to Background Handler module
Added new functions for background handler module which uses a different method for changing backgrounds and handle character transitions to the new backgrounds!
The old functions still exist in the BackgroundHandler module, under the region OBSOLETE. They may be removed someday in the future.
Improved functions from DemoScripter scene script.
Added wait_signal_func for waiting specific signals to be called in order to run specific functions. This is useful in case you want to, for example, run a function when text animation is finished! (Like show a character when text animation finishes, etc)